requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.3.0
autobahn==25.10.2
Automat==25.4.16
bcrypt==4.1.2
beautifulsoup4==4.13.5
bidict==0.23.1
billiard==4.2.2
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.4
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
chardet==3.0.4
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
cssselect==1.3.0
cssselect2==0.8.0
daphne==4.2.1
dateparser==1.2.2
deep-translator==1.11.4
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.6
django-activity-stream==2.0.0
django-allauth==65.13.0
django-celery-beat==2.8.1
django-cleanup==8.0.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-restframework==0.0.1
django-timezone-field==7.1
django-tinymce==5.0.0
django-unfold==0.67.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-yasg==1.21.11
eventlet==0.40.4
fake-useragent==2.2.0
Flask==3.0.0
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-PyMongo==2.3.0
flask-sock==0.7.0
Flask-SocketIO==5.6.0
freetype-py==2.5.1
frozenlist==1.8.0
googletrans==4.0.0rc1
graphviz==0.21
greenlet==3.3.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib_metadata==8.7.0
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
kombu==5.5.4
langdetect==1.0.9
lxml==6.0.1
lxml_html_clean==0.4.2
MarkupSafe==3.0.3
msgpack==1.1.2
multidict==6.7.0
numpy==2.3.3
oauthlib==3.3.1
openai==0.28.1
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
parse==1.20.2
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2==2.9.10
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydotplus==2.0.2
pyee==11.1.1
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.8.0
pymongo==4.15.5
pyOpenSSL==25.3.0
pyparsing==3.2.3
pypdf==6.2.0
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-bidi==0.6.7
python-crontab==3.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.13.0
python-magic==0.4.27
python-magic-bin==0.4.14
python-social-auth==0.3.6
python-socketio==5.16.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==6.4.0
regex==2026.1.15
reportlab==4.4.4
requests==2.32.5
requests-html==0.10.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rfc3986==1.5.0
rlPyCairo==0.4.0
selenium==4.35.0
service-identity==24.2.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
tqdm==4.67.1
transliterate==1.10.2
trio==0.30.0
trio-websocket==0.12.2
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==1.26.20
vine==5.1.0
w3lib==2.3.1
wcwidth==0.2.14
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.4
wsproto==1.2.0
xhtml2pdf==0.2.17
yarl==1.22.0
zipp==3.23.0
zope.interface==8.1
